Schema, UI Changes & Coupon Module Done

+++ database/migrations/2023_09_14_172203_create_order_items_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->foreignId('product_id')->references('id')->on('products');
             $table->integer('quantity');
             $table->string('size')->nullable();
-            $table->double('unit_price', 13, 2);
-            $table->double('total', 13, 2);
+            $table->decimal('unit_price', 13, 2);
+            $table->decimal('total', 13, 2);
             $table->timestamps();
             $table->softDeletes();
         });
